SNMP: added missing casts to int for printf arguments

+++ src/core/snmp/msg_in.c
@@ -1090,7 +1090,7 @@ snmp_pdu_header_check(struct pbuf *p, u16_t ofs, u16_t pdu_len, u16_t *ofs_ret, 
       snmp_inc_snmpingenerrs();
       break;
     default:
-      LWIP_DEBUGF(SNMP_MSG_DEBUG, ("snmp_pdu_header_check(): unknown error_status: %d\n", m_stat->error_status));
+      LWIP_DEBUGF(SNMP_MSG_DEBUG, ("snmp_pdu_header_check(): unknown error_status: %d\n", (int)m_stat->error_status));
       break;
   }
   ofs += (1 + len_octets + len);
